fix: let parcelNode.update use any updater type
Previously the docs said parcelNode.update() would work with any updater type,
but it was only working with straight values. Fixed this by using prepUpdater.

Also moved out the logic to handle updating child nodes to the asChildNode file
which should help keep bundle sizes lower
