+++ packages/dataparcels/src/parcelNode/asNode.js
@@ -6,10 +6,11 @@ import asRaw from '../parcelData/asRaw';
 import ParcelNode from './ParcelNode';
 
 export default (updater: Function): Function => {
-    let fn = (parcelData: ParcelData): ParcelData => {
+    let fn = (parcelData: ParcelData, changeRequest: *): ParcelData => {
         let parcelNode = new ParcelNode();
         parcelNode._parcelData = parcelData;
-        let result = updater(parcelNode);
+        parcelNode._changeRequest = changeRequest;
+        let result = updater(parcelNode, changeRequest);
         if(!(result instanceof ParcelNode)) {
             throw AsNodeReturnNonParcelNodeError();
         }
